the repository messed up everything
I nearly had to rewrite everything because the files that were meant to be in the last commit disappeared. I don't know how this happened. I think it might have to do something with how I ran a deepscan on bleachbit but I honestly dont know how I could've done this. Files are still missing and I'll have to work on that too. This is all horrible.